fix: adding some checks in card and deck application services

+++ Code/src/Application/CardGame/Card/CardRemover.php
@@ -2,6 +2,7 @@
 
 namespace App\Application\CardGame\Card;
 
+use App\Domain\CardGame\Card\CardNotExist;
 use App\Domain\CardGame\Entity\Card;
 use App\Domain\Shared\Bus\Event\EventBus;
 use App\Infrastructure\Symfony\Doctrine\AppEntityManager;
@@ -22,6 +23,9 @@ class CardRemover
     {
         $cardRepository = $this->em->getRepository(Card::class);
         $card = $cardRepository->findOneBy(['id' => $uuid]);
+        if (null === $card) {
+            throw new CardNotExist();
+        }
 
         $card->remove();
         $this->em->remove($card);
